requirements.txt
aniso8601==7.0.0
attrs==19.1.0
bidict==0.23.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dnspython==2.7.0
eventlet==0.38.2
Flask==1.1.1
flask-restx==1.3.0
Flask-SocketIO==5.5.0
greenlet==3.1.1
h11==0.14.0
idna==2.8
importlib_resources==6.4.5
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.0.2
MarkupSafe==1.1.1
pyrsistent==0.15.4
python-engineio==4.11.1
python-socketio==5.12.0
pytz==2019.2
requests==2.25.1
RPi.GPIO==0.7.1
simple-websocket==1.1.0
six==1.12.0
urllib3==1.26.5
Werkzeug==0.15.5
wsproto==1.2.0
